requirements.txt
ccxt
python-dotenv
python-telegram-bot>=20.0a4
